Fixes for PR. Set default --item-size of icons to 48.
