+++ src/app/inventory/ItemRating.tsx
@@ -1,26 +1,21 @@
-import * as React from "react";
-import { D2Item } from "./item-types";
-import showRatingFn from "./show-ratings";
-import "./ItemRender.scss";
-import { dtrRatingColor } from "../shell/dimAngularFilters.filter";
+import * as React from 'react';
+import { dtrRatingColor } from '../shell/dimAngularFilters.filter';
+import './ItemRender.scss';
 
 interface Props {
-  item: D2Item;
+  rating?: number;
+  hideRating?: boolean;
 }
 
 export default class ItemRender extends React.Component<Props> {
   render() {
-    const { item } = this.props;
-    const showRating = showRatingFn(item);
+    const { rating, hideRating } = this.props;
 
-    if (showRating && item.dtrRating && item.dtrRating.overallScore) {
+    if (rating !== undefined && !hideRating) {
       return (
         <div className="review-score">
-          {item.dtrRating.overallScore}
-          <i
-            className="fa fa-star"
-            style={dtrRatingColor(item.dtrRating.overallScore)}
-          />
+          {rating}
+          <i className="fa fa-star" style={dtrRatingColor(rating)} />
         </div>
       );
     } else {
