+++ src/app/inventory/ItemRender.tsx
@@ -1,34 +1,23 @@
 import * as React from 'react';
-import { D2Item } from './item-types';
 import classNames from 'classnames';
-import { DestinySocketCategoryStyle } from 'bungie-api-ts/destiny2';
-import ItemRating from './ItemRating';
 import ItemMod from './ItemMod';
+import ItemRating from './ItemRating';
+import { BadgeInfo } from './get-badge-info';
+import { bungieBackgroundStyle } from '../dim-ui/BungieImage';
+import { D2Item } from './item-types';
+import { DestinySocketCategoryStyle } from 'bungie-api-ts/destiny2';
 import './ItemRender.scss';
 
 interface Props {
   item: D2Item;
-  badge: any;
+  badge: BadgeInfo;
+  rating?: number;
+  hideRating?: boolean;
 }
 
 export default class ItemRender extends React.Component<Props> {
   render() {
-    const { item, badge } = this.props;
-
-    const styles = {
-      backgroundImage: `url('https://www.bungie.net${item.icon}')`
-    };
-
-    let className = `item-render`;
-
-    if (item.masterwork) {
-      className = className + ' masterwork';
-    }
-
-    if (item.isExotic) {
-      className = className + ' exotic';
-    }
-
+    const { item, badge, rating, hideRating } = this.props;
     const category =
       item.sockets &&
       item.sockets.categories.find(
@@ -37,13 +26,13 @@ export default class ItemRender extends React.Component<Props> {
 
     return (
       <div
-        className={classNames(`item-render ${item.dmg ? item.dmg : ''}`, {
+        className={classNames(`item-render`, item.dmg || '', {
           masterwork: item.masterwork,
           exotic: item.isExotic
         })}
       >
         <div className="image">
-          <div className="image-well" style={styles} />
+          <div className="image-well" style={bungieBackgroundStyle(item.icon)} />
           <div className="overlay" />
         </div>
         <div className="plugs">
@@ -65,7 +54,7 @@ export default class ItemRender extends React.Component<Props> {
         </div>
         <div className="attributes">
           <div className="area-overlap attribute-1">
-            <ItemRating item={item} />
+            <ItemRating rating={rating} hideRating={hideRating} />
           </div>
           <div className="attribute-2">
             {badge.showBadge && <div className="power">{badge.badgeCount}</div>}
