+++ src/app/inventory/get-badge-info.ts
@@ -1,0 +1,59 @@
+import { DimItem } from './item-types';
+
+export interface BadgeInfo {
+  showBadge: boolean;
+  badgeClassNames: {
+    [key: string]: boolean;
+  };
+  badgeCount: string;
+}
+
+export default function getBadgeInfo(item: DimItem): BadgeInfo {
+  if (!item.primStat && item.objectives) {
+    return processBounty(item);
+  } else if (item.maxStackSize > 1) {
+    return processStackable(item);
+  } else {
+    return processItem(item);
+  }
+}
+
+function processBounty(item: DimItem) {
+  const showBountyPercentage = !item.complete && !item.hidePercentage;
+
+  const result = {
+    showBadge: showBountyPercentage,
+    badgeClassNames: {},
+    badgeCount: ''
+  };
+
+  if (showBountyPercentage) {
+    result.badgeClassNames = { 'item-stat': true, 'item-bounty': true };
+    result.badgeCount = `${Math.floor(100 * item.percentComplete)}%`;
+  }
+
+  return result;
+}
+
+function processStackable(item: DimItem) {
+  return {
+    showBadge: true,
+    badgeClassNames: { 'item-stat': true, 'item-stackable-max': item.amount === item.maxStackSize },
+    badgeCount: item.amount.toString()
+  };
+}
+
+function processItem(item: DimItem) {
+  const result = {
+    showBadge: Boolean(item.primStat && item.primStat.value),
+    badgeClassNames: {
+      'item-equipment': true
+    },
+    badgeCount: ''
+  };
+  if (item.primStat && result.showBadge) {
+    result.badgeClassNames['item-stat'] = true;
+    result.badgeCount = item.primStat.value.toString();
+  }
+  return result;
+}
