+++ src/app/settings/settings.component.ts
@@ -155,7 +155,7 @@ export function SettingsController(
   vm.resetHiddenInfos = resetHiddenInfos;
 
   vm.resetItemSize = () => {
-    vm.settings.itemSize = window.matchMedia('(max-width: 1025px)').matches ? 38 : 44;
+    vm.settings.itemSize = window.matchMedia('(max-width: 1025px)').matches ? 38 : 48;
   };
 
   vm.saveAndReloadReviews = () => {
